\name{acutesampled.loglkl}
\alias{acutesampled.loglkl}
\title{Compute the log likelihood of a tip cluster table under the \code{Acute} model}
\usage{
  acutesampled.loglkl(tpc, rate.m, sample.prob, dT,
    lclu.n = NULL)
}
\description{
  Compute the log likelihood of a tip cluster table under
  the \code{Acute} model
}

