\name{birthsampled.loglkl}
\alias{birthsampled.loglkl}
\title{Compute the log likelihood of a tip cluster table under the \code{Acute} model}
\usage{
  birthsampled.loglkl(tpc, rate, sample.prob, dT)
}
\description{
  Compute the log likelihood of a tip cluster table under
  the \code{Acute} model
}

