\name{clu.n.of.tchain}
\alias{clu.n.of.tchain}
\title{For each tip cluster with \code{n} transmissions (col) return the number of spanning trees that correspond to \code{i} transmissions from the index case (row)}
\usage{
  clu.n.of.tchain(max.ntransm)
}
\description{
  For each tip cluster with \code{n} transmissions (col)
  return the number of spanning trees that correspond to
  \code{i} transmissions from the index case (row)
}

