\name{clu.sample}
\alias{clu.sample}
\title{For a complete tip cluster contingency table \code{clu}, compute a downsampled version based on sampling probabilities that are random conditional on dependent variables}
\usage{
  clu.sample(clu, s, mx.s.ntr = ncol(clu) - 1, rtn.exp = 0)
}
\description{
  For a complete tip cluster contingency table \code{clu},
  compute a downsampled version based on sampling
  probabilities that are random conditional on dependent
  variables
}

