\name{clu.subtrees.find}
\alias{clu.subtrees.find}
\title{For a given number \code{nIdx} of free slots, compute all possible combinations of subtrees whose total size is \code{nE}.}
\usage{
  clu.subtrees.find(nIdx, nE)
}
\description{
  For a given number \code{nIdx} of free slots, compute all
  possible combinations of subtrees whose total size is
  \code{nE}.
}

