\name{clu.tipc.n}
\alias{clu.tipc.n}
\title{For each tip cluster with \code{n} transmissions (col) return the number of spanning trees that correspond to \code{i} transmissions from the index case (row)}
\usage{
  clu.tipc.n(closure)
}
\description{
  For each tip cluster with \code{n} transmissions (col)
  return the number of spanning trees that correspond to
  \code{i} transmissions from the index case (row)
}

