\name{phdes.plot.confint}
\alias{phdes.plot.confint}
\title{plot binomial confidence intervals}
\usage{
  phdes.plot.confint(p0, p1, c0, c1, xlab = "s.consent",
    ylab = "conf.int", cols = c("red", "blue"),
    legend.loc = "topright", legend.txt = c("", ""))
}
\description{
  plot binomial confidence intervals
}

