\name{popart.get.sampled.acute2acute.p}
\alias{popart.get.sampled.acute2acute.p}
\title{compute the number of acute to acute transmissions from tip clusters up to order 3
this is replaced by "popart.get.sampled.acute2acute"}
\usage{
  popart.get.sampled.acute2acute.p(tipc.p, pop.size, f.coh,
    f.susc, f.untreated, cohort.dur, perc.inc, p.lab,
    p.nocontam, p.consent.coh, p.consent.clu,
    p.prev.instudy.clu, p.inc.instudy.clu,
    opt.power = "All")
}
\description{
  compute the number of acute to acute transmissions from
  tip clusters up to order 3 this is replaced by
  "popart.get.sampled.acute2acute"
}

