\name{popart.get.sampled.transmissions.p}
\alias{popart.get.sampled.transmissions.p}
\title{compute the fraction of linked incident sequences based on overall proportions
this is replaced by "popart.get.sampled.transmissions"}
\usage{
  popart.get.sampled.transmissions.p(pop.size, f.coh,
    f.susc, f.untreated, cohort.dur, perc.inc, p.lab,
    p.nocontam, p.consent.coh, p.consent.clu,
    p.prev.instudy.clu, p.inc.instudy.clu,
    opt.power = "All")
}
\description{
  compute the fraction of linked incident sequences based
  on overall proportions this is replaced by
  "popart.get.sampled.transmissions"
}

