\name{popart.getdata.random.arm}
\alias{popart.getdata.random.arm}
\title{randomize allocation of arms to triplets}
\usage{
  popart.getdata.random.arm(randomize.n)
}
\arguments{
  \item{randomize.n}{number of times the randomization is
  repeated}
}
\value{
  data frame of trial sites with allocated arms
}
\description{
  randomize allocation of arms to triplets
}

