\name{popart.getsampled.linkedseq}
\alias{popart.getsampled.linkedseq}
\title{compute the fraction of linked incident sequences based on overall proportions
this is replaced by "popart.getsampled.linkedseq.n"}
\usage{
  popart.getsampled.linkedseq(pop.size, f.coh, f.susc,
    f.untreated, cohort.dur, perc.inc, p.lab, p.nocontam,
    p.consent.coh, p.consent.clu, p.prev.instudy.clu,
    p.inc.instudy.clu, opt.power = "All")
}
\description{
  compute the fraction of linked incident sequences based
  on overall proportions this is replaced by
  "popart.getsampled.linkedseq.n"
}

