\name{tpc.collapse}
\alias{tpc.collapse}
\title{The simulated set of tip clusters contains NA rows to speed up memory management. This function deletes NA rows for all tip clusters}
\usage{
  tpc.collapse(tpc)
}
\description{
  The simulated set of tip clusters contains NA rows to
  speed up memory management. This function deletes NA rows
  for all tip clusters
}

