\name{tpc.tabulate}
\alias{tpc.tabulate}
\title{The list of tip clusters \code{tpc[["trees"]]} contains for each list item "nodes" and "edges".
This function tabulates the number of nodes per dependent variable in "status" for each tip cluster.
It is assumed that missing NA rows have been deleted.}
\usage{
  tpc.tabulate(tpc)
}
\description{
  The list of tip clusters \code{tpc[["trees"]]} contains
  for each list item "nodes" and "edges". This function
  tabulates the number of nodes per dependent variable in
  "status" for each tip cluster. It is assumed that missing
  NA rows have been deleted.
}

